src/requirements/requirements.txt
# pip freeze > requirements.txt
discord.py==1.7.3
Flask==2.0.3
Flask-RESTful==0.3.9
pymongo==4.0.1
requests==2.27.1